- improve delete warnings - order contacts and staff by name - fix job form bug - move seed data method to factories
