+++ auto_repair_saas/apps/utils/factories.py
@@ -1,4 +1,7 @@
 import factory
+from django.contrib.auth.decorators import login_required
+from django.shortcuts import redirect
+from django.urls import reverse
 from faker import Faker
 
 from auto_repair_saas.apps.authentication.models import User
@@ -71,3 +74,29 @@ class JobFactory(factory.django.DjangoModelFactory):
     payment_registered_on = factory.LazyAttribute(
         lambda _: fake.date_between(start_date='-1y', end_date='today')
     )
+
+
+@login_required
+def seed_data(request, *args, **kwargs):
+    # clean tables
+    for model in (Job, Staff, Vehicle, Contact,):
+        model.objects.all().delete()
+
+    for _ in range(0, 20):
+        client = ContactFactory(contact_type='client')
+        vehicle = VehicleFactory(owner=client)
+        JobFactory(client=client, vehicle=vehicle)
+
+    for _ in range(0, 5):
+        ContactFactory(contact_type='supplier')
+
+    # cleanup estimates
+    pending_estimates = Job.objects.filter(status='pending')
+    for estimate in pending_estimates:
+        estimate.paid = False
+        estimate.payment_registered_on = None
+        estimate.save()
+    for job in Job.objects.all().exclude(status='pending'):
+        if not job.paid:
+            job.payment_registered_on = None
+    return redirect(reverse('dashboard'))
