+++ auto_repair_saas/apps/vehicles/models.py
@@ -6,7 +6,7 @@ from auto_repair_saas.apps.utils.models import BaseModel, ModelManager
 
 class Vehicle(BaseModel):
     number_plate = models.CharField(max_length=15)
-    owner = models.ForeignKey(Contact, on_delete=models.RESTRICT)
+    owner = models.ForeignKey(Contact, on_delete=models.CASCADE)
 
     def __str__(self):
         return self.number_plate
